# internal/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /v2/server/{server_id}/autoboards:
    get:
      description: Returns all autoboards for a server with post/refresh counts and
        limit.
      parameters:
      - description: Server ID
        in: path
        name: server_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get server autoboards
      tags:
      - Server Autoboards
    post:
      consumes:
      - application/json
      description: Creates a new autoboard. Fails if the server autoboard limit is
        reached.
      parameters:
      - description: Server ID
        in: path
        name: server_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create an autoboard
      tags:
      - Server Autoboards
  /v2/server/{server_id}/autoboards/{autoboard_id}:
    delete:
      description: Deletes an autoboard by ID.
      parameters:
      - description: Server ID
        in: path
        name: server_id
        required: true
        type: integer
      - description: Autoboard ID
        in: path
        name: autoboard_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete an autoboard
      tags:
      - Server Autoboards
    patch:
      consumes:
      - application/json
      description: Updates fields of an existing autoboard by ID.
      parameters:
      - description: Server ID
        in: path
        name: server_id
        required: true
        type: integer
      - description: Autoboard ID
        in: path
        name: autoboard_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update an autoboard
      tags:
      - Server Autoboards
  /v2/server/{server_id}/bans:
    get:
      description: Returns all banned players for a server.
      parameters:
      - description: Server ID
        in: path
        name: server_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get server bans
      tags:
      - Server Bans
  /v2/server/{server_id}/bans/{player_tag}:
    delete:
      description: Removes a player ban from the server.
      parameters:
      - description: Server ID
        in: path
        name: server_id
        required: true
        type: integer
      - description: Player Tag
        in: path
        name: player_tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Remove a ban
      tags:
      - Server Bans
    post:
      consumes:
      - application/json
      description: Bans a player on the server.
      parameters:
      - description: Server ID
        in: path
        name: server_id
        required: true
        type: integer
      - description: Player Tag
        in: path
        name: player_tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add or update a ban
      tags:
      - Server Bans
...
swagger: "2.0"